Issue 313 - cleanup pagination
Cleanup pagination code, and combine paginationPrep and Pagination
layout

Fixes: #313
