+++ src/pages/PreservedUsers/PreservedUsers.tsx
@@ -28,7 +28,7 @@ import SearchInputLayout from "src/components/layouts/SearchInputLayout";
 // Tables
 import UsersTable from "../../components/tables/UsersTable";
 // Components
-import PaginationPrep from "src/components/PaginationPrep";
+import PaginationLayout from "src/components/layouts/PaginationLayout";
 import BulkSelectorUsersPrep from "src/components/BulkSelectorUsersPrep";
 // Modals
 import DeleteUsers from "src/components/modals/DeleteUsers";
@@ -102,6 +102,7 @@ const PreservedUsers = () => {
     if (userDataResponse.isFetching) {
       setShowTableRows(false);
       // Reset selected users on refresh
+      setUsersTotalCount(0);
       setSelectedUserNames([]);
       setSelectedUserIds([]);
       setSelectedUsers([]);
@@ -153,6 +154,7 @@ const PreservedUsers = () => {
     setShowTableRows(false);
 
     // Reset selected users on refresh
+    setUsersTotalCount(0);
     setSelectedUserNames([]);
     setSelectedUserIds([]);
     setSelectedUsers([]);
@@ -225,6 +227,7 @@ const PreservedUsers = () => {
   // Issue a search using a specific search value
   const submitSearchValue = () => {
     setShowTableRows(false);
+    setUsersTotalCount(0);
     setSearchIsDisabled(true);
     retrieveUser({
       searchValue: searchValue,
@@ -330,13 +333,12 @@ const PreservedUsers = () => {
   };
 
   // Data wrappers
-  // - 'PaginationPrep'
+  // - 'PaginationLayout'
   const paginationData = {
     page,
     perPage,
     updatePage,
     updatePerPage,
-    showTableRows,
     updateSelectedPerPage,
     updateShownElementsList: updateShownUsersList,
     totalCount,
@@ -496,7 +498,7 @@ const PreservedUsers = () => {
     {
       key: 9,
       element: (
-        <PaginationPrep
+        <PaginationLayout
           list={preservedUsersList}
           paginationData={paginationData}
           widgetId="pagination-options-menu-top"
@@ -548,12 +550,11 @@ const PreservedUsers = () => {
             </InnerScrollContainer>
           </OuterScrollContainer>
         </div>
-        <PaginationPrep
+        <PaginationLayout
           list={preservedUsersList}
           paginationData={paginationData}
           variant={PaginationVariant.bottom}
           widgetId="pagination-options-menu-bottom"
-          perPageComponent="button"
           className="pf-v5-u-pb-0 pf-v5-u-pr-md"
         />
       </PageSection>
